restored fadable mixin
